created hidden units generator and for loop to create model
Different models are now creatured via a for loop that is based on the number of hidden units we are at in our hidden units vector.
Hidden units vector is populated with powers of 2. Currently it only holds the first two
Model outputs are saved into a list that we will be able to use for further computation later